Hack mini_magick crash with doc files in parallel
This is a hack around a bug in `identify`, from ImageMagick, that causes
a crash when two Microsoft Office files, like two different Word
documents, are processed at the same time.

If the files are processed one at a time, it works correctly, but if
done at the same time, it crashes.

To fix, MiniMagick::Image.read is patched to wait on a global Mutex when
called for matching file extensions.